Update onStop method override
Now using the custom releaseMediaPlayer method instead of normal release
method. When the onStop method is executed, media player will be checked
to see if it is not null (not initalized); if it is not null then it
will be released and set to null to indicate that the media player is
not initialized to play an audio file at the moment.